Add Readme and Folder Projet: It contains the algorithms of DT, LR, SVM and MP, in addition to the presentation of the operation.
