+++ Unit_4/Project/DT.scala
@@ -1,0 +1,71 @@
+//Import Libraries
+import org.apache.spark.ml.Pipeline
+import org.apache.spark.ml.classification.DecisionTreeClassificationModel
+import org.apache.spark.ml.classification.DecisionTreeClassifier
+import org.apache.spark.ml.evaluation.MulticlassClassificationEvaluator
+import org.apache.spark.ml.feature.{IndexToString, StringIndexer, VectorIndexer}
+import org.apache.spark.ml.feature.VectorAssembler
+import org.apache.spark.sql.SparkSession
+
+//Error level code.
+import org.apache.log4j._
+Logger.getLogger("org").setLevel(Level.ERROR)
+
+//Spark session
+val spark = SparkSession.builder().getOrCreate()
+
+//The data from the bank-full.csv dataset is imported
+val df  = spark.read.option("header","true").option("inferSchema", "true").option("delimiter",";").format("csv").load("bank-full.csv")
+
+//The makes the process of categorizing string variables to numeric
+val yes = df.withColumn("y",when(col("y").equalTo("yes"),1).otherwise(col("y")))
+val cl = yes.withColumn("y",when(col("y").equalTo("no"),2).otherwise(col("y")))
+
+// An array of the selected fields is created in assembler
+val assembler = new VectorAssembler().setInputCols(Array("age","balance","day","duration","campaign","pdays","previous")).setOutputCol("features")
+
+//Transforms into a new dataframe
+val df2 = assembler.transform(df)
+
+//The column and label are given a new name
+val featuresLabel = df2.withColumnRenamed("y", "label")
+
+//Select the indexes
+val dataIndexed = featuresLabel.select("label","features")
+
+
+//Creation of labelIndexer and featureIndexer for the pipeline, Where features with distinct values > 4, are treated as continuous.
+val labelIndexer = new StringIndexer().setInputCol("label").setOutputCol("indexedLabel").fit(dataIndexed)
+val featureIndexer = new VectorIndexer().setInputCol("features").setOutputCol("indexedFeatures").setMaxCategories(4).fit(dataIndexed)
+
+//Training data as 70% and test data as 30%.
+val Array(training, test) = dataIndexed.randomSplit(Array(0.7, 0.3))
+
+//Creating the Decision Tree object.
+val dt = new DecisionTreeClassifier().setLabelCol("indexedLabel").setFeaturesCol("indexedFeatures")
+
+//Creating the Index to String object.
+val labelConverter = new IndexToString().setInputCol("prediction").setOutputCol("predictedLabel").setLabels(labelIndexer.labels)
+
+//Creating the pipeline with the objects created before.
+val pipeline = new Pipeline().setStages(Array(labelIndexer, featureIndexer, dt, labelConverter))
+
+//Fitting the model with training data.
+val model = pipeline.fit(training)
+
+//Making the predictions transforming the testData.
+val predictions = model.transform(test)
+
+//Showing the predictions
+predictions.select("predictedLabel", "label", "features").show(5)
+
+////Creating the evaluator of prediction
+val evaluator = new MulticlassClassificationEvaluator().setLabelCol("indexedLabel").setPredictionCol("prediction").setMetricName("accuracy")
+
+//Accuracy and Test Error
+val accuracy = evaluator.evaluate(predictions)
+println(s"Test Error = ${(1.0 - accuracy)}")
+
+//Show Decision Tree Model
+val treeModel = model.stages(2).asInstanceOf[DecisionTreeClassificationModel]
+println(s"Learned classification tree model:\n ${treeModel.toDebugString}")
